add users

+++ backend/controllers/label.go
@@ -1,7 +1,7 @@
 package controllers
 
 import (
-	"encoding/json"
+	"github.com/vova/pa2020/backend/middleware"
 	"net/http"
 
 	dbpkg "github.com/vova/pa2020/backend/db"
@@ -12,89 +12,13 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
+// todo: vote
 func GetLabels(c *gin.Context) {
-	ver, err := version.New(c)
-	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
-		return
-	}
-
 	db := dbpkg.DBInstance(c)
-	parameter, err := dbpkg.NewParameter(c, models.Label{})
-	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
-		return
-	}
-
-	db, err = parameter.Paginate(db)
-	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
-		return
-	}
-
-	db = parameter.SetPreloads(db)
-	db = parameter.SortRecords(db)
-	db = parameter.FilterFields(db)
-	labels := []models.Label{}
-	fields := helper.ParseFields(c.DefaultQuery("fields", "*"))
-	queryFields := helper.QueryFields(models.Label{}, fields)
-
-	if err := db.Select(queryFields).Find(&labels).Error; err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
-		return
-	}
-
-	index := 0
-
-	if len(labels) > 0 {
-		index = int(labels[len(labels)-1].ID)
-	}
-
-	if err := parameter.SetHeaderLink(c, index); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
-		return
-	}
-
-	if version.Range("1.0.0", "<=", ver) && version.Range(ver, "<", "2.0.0") {
-		// conditional branch by version.
-		// 1.0.0 <= this version < 2.0.0 !!
-	}
-
-	if _, ok := c.GetQuery("stream"); ok {
-		enc := json.NewEncoder(c.Writer)
-		c.Status(200)
-
-		for _, label := range labels {
-			fieldMap, err := helper.FieldToMap(label, fields)
-			if err != nil {
-				c.JSON(400, gin.H{"error": err.Error()})
-				return
-			}
-
-			if err := enc.Encode(fieldMap); err != nil {
-				c.JSON(400, gin.H{"error": err.Error()})
-				return
-			}
-		}
-	} else {
-		fieldMaps := []map[string]interface{}{}
-
-		for _, label := range labels {
-			fieldMap, err := helper.FieldToMap(label, fields)
-			if err != nil {
-				c.JSON(400, gin.H{"error": err.Error()})
-				return
-			}
-
-			fieldMaps = append(fieldMaps, fieldMap)
-		}
-
-		if _, ok := c.GetQuery("pretty"); ok {
-			c.IndentedJSON(200, fieldMaps)
-		} else {
-			c.JSON(200, fieldMaps)
-		}
-	}
+	var labels []models.Label
+	rawQuery := "SELECT * FROM labels where user_id = ?;"
+	db.Raw(rawQuery, middleware.UserInstance(c).ID).Scan(&labels)
+	c.JSON(201, labels)
 }
 
 func GetLabel(c *gin.Context) {
@@ -123,7 +47,14 @@ func GetLabel(c *gin.Context) {
 		return
 	}
 
+	if !vote(c, label) {
+		c.JSON(403, "you dont have permissions")
+		c.Abort()
+		return
+	}
+
 	fieldMap, err := helper.FieldToMap(label, fields)
+
 	if err != nil {
 		c.JSON(400, gin.H{"error": err.Error()})
 		return
@@ -156,6 +87,12 @@ func CreateLabel(c *gin.Context) {
 		return
 	}
 
+	if !vote(c, label) {
+		c.JSON(403, "you dont have permissions")
+		c.Abort()
+		return
+	}
+
 	if err := db.Create(&label).Error; err != nil {
 		c.JSON(400, gin.H{"error": err.Error()})
 		return
@@ -185,6 +122,12 @@ func UpdateLabel(c *gin.Context) {
 		return
 	}
 
+	if !vote(c, label) {
+		c.JSON(403, "you dont have permissions")
+		c.Abort()
+		return
+	}
+
 	if err := db.Save(&label).Error; err != nil {
 		c.JSON(400, gin.H{"error": err.Error()})
 		return
@@ -194,12 +137,6 @@ func UpdateLabel(c *gin.Context) {
 }
 
 func DeleteLabel(c *gin.Context) {
-	ver, err := version.New(c)
-	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
-		return
-	}
-
 	db := dbpkg.DBInstance(c)
 	id := c.Params.ByName("id")
 	label := models.Label{}
@@ -209,15 +146,15 @@ func DeleteLabel(c *gin.Context) {
 		c.JSON(404, content)
 		return
 	}
+	if !vote(c, label) {
+		c.JSON(403, "you dont have permissions")
+		c.Abort()
+		return
+	}
 
 	if err := db.Delete(&label).Error; err != nil {
 		c.JSON(400, gin.H{"error": err.Error()})
 		return
-	}
-
-	if version.Range("1.0.0", "<=", ver) && version.Range(ver, "<", "2.0.0") {
-		// conditional branch by version.
-		// 1.0.0 <= this version < 2.0.0 !!
 	}
 
 	c.Writer.WriteHeader(http.StatusNoContent)
